codigos y respuestas metodos PUT, PATCH y DELETE en la entidad propiedad

+++ src/propiedad/propiedad.service.ts
@@ -3,14 +3,14 @@ import { Propiedad } from './propiedad.interface';
 // Firebase:
 import { collection, getDocs, Firestore, DocumentData, addDoc, updateDoc, doc, setDoc, deleteDoc, getDoc } from 'firebase/firestore/lite';
 import { error } from 'console';
-
+import db from 'src/firebase/config';
 
 
 
 @Injectable()
 export class PropiedadService {
 
-    async getPropiedades(db: Firestore): Promise<any[]> {
+    async getPropiedades(): Promise<any[]> {
         //referencia a la base de datos con la coleccion/array "propiedades"
         const propiedadCollection = collection(db, 'propiedades');
         // traeme el collecion de la referencia creada anteriormente y guardale en una variable.
@@ -32,7 +32,7 @@ export class PropiedadService {
     }
 
 
-    async getPropiedadById(db: Firestore, id: string): Promise<DocumentData> {
+    async getPropiedadById(id: string): Promise<DocumentData> {
 
         const refDoc = doc(db, "propiedades", id);
         const docPorId = await getDoc(refDoc);
@@ -47,49 +47,53 @@ export class PropiedadService {
         } else return new error()
     }
 
-    async createPropiedad(db: Firestore, propiedad: any):Promise<DocumentData> {
+    async createPropiedad(propiedad: any): Promise<DocumentData> {
         try {
             await addDoc(collection(db, "propiedades"), propiedad);
-        return (
+            return (
                 {
-                 ...propiedad
+                    ...propiedad
                 }
-        );
-        }catch (error){
+            );
+        } catch (error) {
             throw new Error("created failed")
         }
     }
 
-    
-
-    async patchPropiedad(db: Firestore, body: any, id: string) {
-
-        const docRef = doc(db, "propiedades",
-            id);
-        // update actualiza solamente el campo que le mandemos en el body, dejando los otros campos como estaban.
-        await updateDoc(docRef, body);
-
+    async patchPropiedad(body: any, id: string):Promise<{ success: boolean, message: string }> {
+        const docRef = doc(db, "propiedades", id);
+        const docPorId = await getDoc(docRef);
+        if (docPorId.exists()) {
+            // update actualiza solamente el campo que le mandemos en el body, dejando los otros campos como estaban.
+            await updateDoc(docRef, body);
+            return { success: true, message: `Propiedad edited` }
+        } else {
+            return new error()
+        }
     }
 
-    async putPropiedad(db: Firestore, body: any, id: string) {
-
-        const docRef = doc(db, "propiedades",
-            id);
-
-        // setdoc es destructivo, osea que remplaza el body del docuento por el body que le mandemos.
-        await setDoc(docRef, body);
-
+    async putPropiedad(body: any, id: string): Promise<{ success: boolean, message: string }> {
+        const docRef = doc(db, "propiedades", id);
+        const docPorId = await getDoc(docRef);
+        if (docPorId.exists()) {
+            // setdoc es destructivo, osea que remplaza el body del docuento por el body que le mandemos.
+            await setDoc(docRef, body);
+            return { success: true, message: `Propiedad edited` }
+        } else {
+            return new error()
+        }
     }
 
-    async deletePropiedad(db: Firestore, id: string) {
+    async deletePropiedad(id: string): Promise<{ success: boolean, message: string }> {
+        const docRef = doc(db, "propiedades", id);
+        const docPorId = await getDoc(docRef);
+        if (docPorId.exists()) {
+            await deleteDoc(docRef);
+            return { success: true, message: `Propiedad with id: {${id}} was deleted` }
 
-        const docRef = doc(db, "propiedades",
-            id);
-
-        await deleteDoc(docRef);
-
+        } else {
+            return new error()
+        }
     }
-
-
 
 }
